requirements.txt
fastapi
uvicorn
requests
tweepy
python-dotenv